+++ src/components/details-info-about-offer/details-info-about-offer.jsx
@@ -15,7 +15,7 @@ const DetailsInfoAboutOffer = (props) => {
   const {offer, onTitleClick} = props;
   const {title, images, price, type, rating, bedrooms, maxAdults, goods, isFavorite, isPremium, host, description} = offer;
   const {avatarUrl, name} = host;
-  const additionalClass = OfferPrefix.DETAILS_PAGE;
+  const additionalClass = OfferPrefix.DETAILS;
   const nearbyOffers = offers.slice(0, MAX_NEIGHBOURHOOD_OFFERS);
   const isNearbyOffers = nearbyOffers.length !== 0;
 
@@ -106,10 +106,10 @@ const DetailsInfoAboutOffer = (props) => {
               <div className="property__inside">
                 <h2 className="property__inside-title">What&rsquo;s inside</h2>
                 <ul className="property__inside-list">
-                  {goods.map((good) => {
+                  {goods.map((it) => {
                     return (
-                      <li key={`${good}`} className="property__inside-item">
-                        good
+                      <li key={`${it}`} className="property__inside-item">
+                        {it}
                       </li>
                     );
                   })}
