+++ src/components/offer-neighbourhood/offer-neighbourhood.test.js
@@ -5,7 +5,7 @@ import {offers} from "../../mocks/test-offer";
 import {MAX_NEIGHBOURHOOD_OFFERS, OfferPrefix} from "../../const";
 
 it(`Should render <OfferNeighbourhood/>`, () => {
-  const tree = renderer.create(<OfferNeighbourhood additionalClass={OfferPrefix.DETAILS_PAGE} renderOffer={() => {}} offers={offers.slice(0, MAX_NEIGHBOURHOOD_OFFERS)} onTitleClick={() => {
+  const tree = renderer.create(<OfferNeighbourhood additionalClass={OfferPrefix.DETAILS} renderOffer={() => {}} offers={offers.slice(0, MAX_NEIGHBOURHOOD_OFFERS)} onTitleClick={() => {
   }}/>).toJSON();
 
   expect(tree).toMatchSnapshot();
