+++ src/const.js
@@ -2,9 +2,24 @@ const ZOOM = 12;
 
 const MAX_NEIGHBOURHOOD_OFFERS = 3;
 
-const OfferPrefix = {
-  MAIN_PAGE: `cities__place-card`,
-  DETAILS_PAGE: `near-places__card`,
+const monthMap = {
+  "0": `January`,
+  "1": `February`,
+  "2": `March`,
+  "3": `April`,
+  "4": `May`,
+  "5": `June`,
+  "6": `July`,
+  "7": `August`,
+  "8": `September`,
+  "9": `October`,
+  "10": `November`,
+  "11": `December`,
 };
 
-export {ZOOM, MAX_NEIGHBOURHOOD_OFFERS, OfferPrefix};
+const OfferPrefix = {
+  MAIN: `cities__place-card`,
+  DETAILS: `near-places__card`,
+};
+
+export {ZOOM, MAX_NEIGHBOURHOOD_OFFERS, OfferPrefix, monthMap};
