+++ src/utils/index.js
@@ -1,23 +1,4 @@
-const monthMap = {
-  "0": `January`,
-  "1": `February`,
-  "2": `March`,
-  "3": `April`,
-  "4": `May`,
-  "5": `June`,
-  "6": `July`,
-  "7": `August`,
-  "8": `September`,
-  "9": `October`,
-  "10": `November`,
-  "11": `December`,
-};
-
-
-const getRandomInteger = (min, max) => {
-  const rand = min + Math.random() * (max + 1 - min);
-  return Math.floor(rand);
-};
+import {monthMap} from "../const";
 
 const formatDate = (date) => {
   const dateFormat = new Date(date);
@@ -37,4 +18,4 @@ const formatDateTime = (date) => {
   return `${year}-${month}-${day}`;
 };
 
-export {getRandomInteger, formatDate, formatDateTime};
+export {formatDate, formatDateTime};
